requirements.txt
torch
scikit-learn







mysql-connector-python
sqlalchemy